init
This takes the binary release of delta.chat, because the source doesn't
compile due to some node madness. Whoever thought that Electron was a
good target for writing apps needs to see a professional.

The app also doesn't start, because... actually I don't know. I can't
strace the process, because it ptraces itself. srsly. wtf.